make sure the return dtype of unravel_index of torch is the default int64

+++ ivy/functional/backends/torch/experimental/statistical.py
@@ -145,7 +145,7 @@ def unravel_index(
     *,
     out: Optional[torch.Tensor] = None,
 ) -> torch.Tensor:
-    temp = indices
+    temp = indices.to("int64")
     output = []
     for dim in reversed(shape):
         output.append(temp % dim)
